Fix split packages in external test modules (#78136)
The external test modules are plugins which provide extra test
functionality in snapshot builds. These are independent Elasticsearch
plugins added as modules. They currently share package names with
server. This commit moves them each into their own java package name.
