+++ test/external-modules/delayed-aggs/build.gradle
@@ -8,17 +8,11 @@
 
 esplugin {
   description 'A test module that allows to delay aggregations on shards with a configurable time'
-  classname 'org.elasticsearch.search.aggregations.DelayedShardAggregationPlugin'
+  classname 'org.elasticsearch.test.delayedshard.DelayedShardAggregationPlugin'
 }
 
 restResources {
   restApi {
     include '_common', 'indices', 'index', 'cluster', 'search'
   }
-}
-
-tasks.named('splitPackagesAudit').configure {
-  // aggs is owned by server, these should be moved to delayedaggs
-  ignoreClasses 'org.elasticsearch.search.aggregations.DelayedShardAggregationBuilder',
-    'org.elasticsearch.search.aggregations.DelayedShardAggregationPlugin'
 }
