+++ test/external-modules/die-with-dignity/build.gradle
@@ -6,7 +6,7 @@ apply plugin: 'elasticsearch.internal-es-plugin'
 
 esplugin {
   description 'Die with dignity plugin'
-  classname 'org.elasticsearch.DieWithDignityPlugin'
+  classname 'org.elasticsearch.test.diewithdignity.DieWithDignityPlugin'
 }
 
 // let the javaRestTest see the classpath of main
@@ -30,10 +30,4 @@ tasks.named("test").configure {
 
 tasks.named("yamlRestTest").configure {
   enabled = false
-}
-
-tasks.named('splitPackagesAudit').configure {
-  // these should be moved to an actual package, not the root package
-  ignoreClasses 'org.elasticsearch.DieWithDignityPlugin',
-    'org.elasticsearch.RestDieWithDignityAction'
 }
