+++ test/external-modules/error-query/build.gradle
@@ -8,18 +8,11 @@
 
 esplugin {
   description 'A test module that exposes a way to simulate search shard failures and warnings'
-  classname 'org.elasticsearch.search.query.ErrorQueryPlugin'
+  classname 'org.elasticsearch.test.errorquery.ErrorQueryPlugin'
 }
 
 restResources {
   restApi {
     include '_common', 'indices', 'index', 'cluster', 'search'
   }
-}
-
-tasks.named('splitPackagesAudit').configure {
-  // search.query is owned by server, these should be moved to errorquery
-  ignoreClasses 'org.elasticsearch.search.query.ErrorQueryBuilder',
-    'org.elasticsearch.search.query.ErrorQueryPlugin',
-    'org.elasticsearch.search.query.IndexError'
 }
