+++ test/external-modules/error-query/src/main/java/org/elasticsearch/test/errorquery/IndexError.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.search.query;
+package org.elasticsearch.test.errorquery;
 
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
